[DF] Generate list of valid branch names once
- GetBranchNames is now a public method of RLoopManager
- fBranchNames is now RLoopManager::fValidBranchNames
- the first call to GetBranchNames evaluates the list of
  valid branch names (expensive operation) and stores it
  in fValidBranchNames. Subsequent calls simply return
  the cached value
- code in RInterface has been simplified as a consequence
